Make the extension builder run webpack directly.
Fixes #9034
